Weapon sway/hands sway
Added some sway to the fps rig when the player turns.
The sway is based on a tutorial by Plai on youtube. Haven't found a good way to introduce some sway that looks nice without parenting the fps rig to the virtual camera, this will have to do for now.