Ray_Spells works
Ray_Spells function like they should now but there's still room for adjustments. Currently the ray will try to aim for the center of each enemy but if the collider is offset from the ground it will aim slightly under center mass. Ray_Spells can be adjusted to use a cone or the entire sphere to check for valid hits, this allows both for directed attacks and aoe attacks as well.