Spell_Targeting V1
Targeting works but isn't optimized and probably needs some more tweaking before merging with the main branch.

Targeting works by highlighting objects, this is done by setting the emissive value in a new shader made for this. A script on the object itself will make a material-property-block and set some values. When targeting, the targeting-script will call a method on the material instance-script setting the emissive value to 1 or higher(the value is clamped in the shader so iit will allways be between 0 and 1), then the material instance script will decrease the emissive value over time and make the glow fade away.