Changed spell casting.
Spells now use a mana system instead of a cooldown system. The player has a limited supply of mana points that regenerate over time and each time they cast a spell it costs some mana. There is a small delay between each cast but it is small enough that it won't be relevant for players to be told about it. A ball with some text on it displays how much mana the player has left. The player now has to hold the left and right mouse buttons to cast spells instead of tapping them.