Test day preparations.
Added a tutorial where players can test their abilities. Added damage over time effect for wind abilities, just like all the other elements. Added a slow down effect that is applied to enemies by ice and earth spells. The navmeshagent does not change speed when told to, this is an error that will need to be fixed later.