Enemy fixes.
Changed bat layer so it can be damaged by spells. Fixed skeletons in tutorial. Made so the dastardly duo in the menues did not appear in front of the info screens. Enemies not getting a spawner and staying on the ground after death does not seem to be much of an issue anymore?