UI Spellbook + UI fixes
Adding a spellbook to the UI that the player can use to look up spells. Might add more info on enemies once the spells are done. Did some fixes to the UI, problems are related to canvas having too low resolution as well as scaling with height instead of width.