Bats ragdoll and small fixes
Bats no have a collider and rigidbody that activates when the die, because the body and the wings are seperate objects connecting them with hinges would be a bit of work. Removed additional audio listener from start scene. fixed up UI a bit so some elements activate/deactivate when buttons are pressed.